pyproject.toml
[tool.poetry]
name = "cargo-insurance"
version = "0.1.0"
description = "Rest API for cargo insurance"
authors = ["Mark Krylov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.68.1"
tortoise-orm = "^0.19.0"
uvicorn = "^0.15.0"
aerich = "^0.7.1"
asyncpg = "^0.28.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"

[tool.black]
line-length = 79
target-version = ["py311"]

[tool.aerich]
tortoise_orm = "src.core.tortoise_conf.TORTOISE_ORM"
location = "./src/db/migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
